[Update] sign in form style
