+++ src/app.js
@@ -5,7 +5,7 @@ import { Provider } from 'react-redux';
 import {
   Router,
   Route,
-  browserHistory
+  hashHistory
 } from 'react-router';
 
 import thunk from 'redux-thunk';
@@ -16,7 +16,7 @@ import reducers from 'flow/reducers';
 
 import App from 'components/app';
 import { About } from 'components/about';
-import Signin from 'components/signin';
+import Signin from 'components/sign-in';
 
 import firebaseService from 'services/firebase';
 window.firebase = firebaseService;
@@ -29,7 +29,7 @@ const store = createStore(reducers, compose(
 
 render((
   <Provider store={ store }>
-    <Router history={ browserHistory }>
+    <Router history={ hashHistory }>
       <Route path='/' component={ App }>
         <Route path='/about' component={ About } />
         <Route path='/signin' component={ Signin } />
