+++ src/components/app/style/app.styl
@@ -1,0 +1,24 @@
+html
+  height: 100%;
+  width: 100%;
+  font-size: remBase;
+
+body
+  background: linear-gradient(to bottom right, baseToneColor, baseColor);
+  font-size: h1Size
+  font-family: sans-serif
+  color: fontColor;
+  margin: 0
+  padding: 0
+  height: 100%;
+
+.app
+  display: flex
+  flex-direction: column
+  text-align: center
+  height: 100%;
+
+  .app-view
+    height: 100%;
+    display: flex
+    flex-direction: column
