+++ src/components/app/style/index.styl
@@ -1,0 +1,1 @@
+@import './app'
