+++ src/components/sign-in/index.js
@@ -1,0 +1,3 @@
+import _Signin from './sign-in';
+
+export default _Signin;
