+++ src/components/sign-in/style/sign-in.styl
@@ -1,0 +1,44 @@
+.sign-in
+  display: flex
+  flex-direction: column
+  background: accentColor
+  align-self: center
+  width: 23rem
+  padding: 3rem
+  height: 50%
+  box-shadow: 0px 0px 10px rgba(0,0,0,.3)
+
+  .sign-in-header
+    display: flex
+    flex-direction: column
+    flex: 1
+
+    .sign-in-welcome-text
+      color: fontColor
+      font-size: 2rem
+      text-align: left
+      margin-bottom: 2rem
+
+    .sign-in-title
+      font-size: 2rem
+      text-align: left
+
+
+      span
+        color: fontSecondaryColor
+
+  .sign-in-form
+    display: flex
+    flex: 1
+    margin-top: 3rem
+    margin-bottom: 3rem
+    flex-direction: column
+    align-items: flex-start
+
+  .sign-in-footer
+    color:fontSecondaryColor
+
+    a
+      color: fontSecondaryColor
+      font-weight: bold
+      text-decoration: none
