--- src/components/signin/index.js
@@ -1,3 +1,0 @@
-import _Signin from './signin';
-
-export default _Signin;
