+++ src/styles/main.styl
@@ -2,18 +2,5 @@
 @import '../components/title/main'
 @import '../components/about/style'
 @import '../components/navigation/style'
-
-body
-  background-color: baseColor
-  font-size: h1Size
-  font-family: sans-serif
-  margin: 0
-  padding: 0
-
-.app
-  display: flex
-  flex-direction: column
-  text-align: center
-
-  .view
-    display: flex
+@import '../components/sign-in/style'
+@import '../components/app/style'
