+++ src/styles/variables.styl
@@ -1,7 +1,10 @@
-baseColor = darkGrey
+baseColor = greyBlue
+baseToneColor = greyBlueLight
 accentColor = white
-fontColor = jordyBlue
+fontColor = darkGrey
+fontSecondaryColor = lightGrey
 
+remBase = 12px
 h1Size = 1rem
 
 white = #ffffff
@@ -9,7 +12,11 @@ green = #b3ddb3
 lightGreen = #d1ffd7
 dargGreen = #a6d3a0
 grey = #808782
-darkGrey = #264348
 purple = #9B59B6
 jordyBlue = #89C4F4
 steelBlue = #4B77BE
+lightGrey = #CFD5DC
+mideGrey = #8F9CAE
+darkGrey = #46505F
+greyBlueLight = #AEC2D4
+greyBlue = #849CB2
